Move symbols into Foundation
Bounds, CompositionAlignment, Padding, Autosize, RepeatFields,
CInterpolate.
