+++ hvega/src/Graphics/Vega/VegaLite/Foundation.hs
@@ -52,6 +52,13 @@ module Graphics.Vega.VegaLite.Foundation
        , Resolve(..)
        , Resolution(..)
 
+       , Bounds(..)
+       , CompositionAlignment(..)
+       , Padding(..)
+       , Autosize(..)
+       , RepeatFields(..)
+       , CInterpolate(..)
+
        -- not for external export
        , fontWeightSpec
        , measurementLabel
@@ -75,6 +82,12 @@ module Graphics.Vega.VegaLite.Foundation
        , ttContentLabel
        , channelLabel
        , resolveProperty
+       , boundsSpec
+       , compositionAlignmentSpec
+       , paddingSpec
+       , autosizeProperty
+       , repeatFieldsProperty
+       , cInterpolateSpec
 
        , fromT
        , field_
@@ -954,3 +967,169 @@ resolveProperty res =
 
       ans = map (\(ch, rule) -> channelLabel ch .= resolutionLabel rule) rls
   in (nme, object ans)
+
+
+-- | This is used with 'Graphics.Vega.VegaLite.bounds' to define the extent of a sub plot.
+--
+--   @since 0.4.0.0
+
+data Bounds
+  = Full
+    -- ^ Bounds calculation should use the entire plot area (including axes, title,
+    --   and legend).
+  | Flush
+    -- ^ Bounds calculation should take only the specified width and height values for
+    --   a sub-view. Useful when attempting to place sub-plots without axes or legends into
+    --   a uniform grid structure.
+
+
+boundsSpec :: Bounds -> VLSpec
+boundsSpec Full = "full"
+boundsSpec Flush = "flush"
+
+
+-- | Specifies the alignment of compositions. It is used with:
+--   'Graphics.Vega.VegaLite.align', 'Graphics.Vega.VegaLite.alignRC', 'Graphics.Vega.VegaLite.LeGridAlign', and 'Graphics.Vega.VegaLite.LGridAlign'.
+--
+--   @since 0.4.0.0
+
+data CompositionAlignment
+    = CANone
+    -- ^ Flow layout is used, where adjacent subviews are placed one after
+    --   another.
+    | CAEach
+    -- ^ Each row and column may be of a variable size.
+    | CAAll
+    -- ^ All the rows and columns are of the same size (this is based on the
+    --   maximum subview size).
+
+
+compositionAlignmentSpec :: CompositionAlignment -> VLSpec
+compositionAlignmentSpec CANone = "none"
+compositionAlignmentSpec CAEach = "each"
+compositionAlignmentSpec CAAll = "all"
+
+
+-- | Specify the padding dimensions in pixel units.
+
+data Padding
+    = PSize Double
+      -- ^ Use the same padding on all four edges of the container.
+    | PEdges Double Double Double Double
+      -- ^ Specify the padding for the left, top, right, and bottom edges.
+
+
+paddingSpec :: Padding -> VLSpec
+paddingSpec (PSize p) = toJSON p
+paddingSpec (PEdges l t r b) =
+  object [ "left" .= l
+         , "top" .= t
+         , "right" .= r
+         , "bottom" .= b
+         ]
+
+
+{-|
+
+Indicates the auto-sizing characteristics of the visualization such as amount
+of padding, whether it should fill the parent container etc. For more details see the
+<https://vega.github.io/vega-lite/docs/size.html#autosize Vega-Lite documentation>.
+
+-}
+
+data Autosize
+    = AContent
+      -- ^ Interpret visualization dimensions to be for the data rectangle (external
+      --   padding added to this size).
+    | AFit
+      -- ^ Interpret visualization dimensions to be for the entire visualization (data
+      --   rectangle is shrunk to accommodate external decorations padding).
+    | ANone
+      -- ^ No autosizing is applied.
+    | APad
+      -- ^ Automatically expand size of visualization from the given dimensions in order
+      --   to fit in all supplementary decorations (legends etc.).
+    | APadding
+      -- ^ Interpret visualization width to be for the entire visualization (data
+      -- rectangle is shrunk to accommodate external padding).
+    | AResize
+      -- ^ Recalculate autosizing on every view update.
+
+
+autosizeProperty :: Autosize -> LabelledSpec
+autosizeProperty APad = ("type", fromT "pad")
+autosizeProperty AFit = ("type", fromT "fit")
+autosizeProperty ANone = ("type", fromT "none")
+autosizeProperty AResize = "resize" .= True
+autosizeProperty AContent = ("contains", fromT "content")
+autosizeProperty APadding = ("contains", fromT "padding")
+
+
+{-|
+
+Create a list of fields to use in set of repeated small multiples. The list of
+fields named here can be referenced in an encoding with @'Graphics.Vega.VegaLite.PRepeat' 'Graphics.Vega.VegaLite.Column'@
+or @'Graphics.Vega.VegaLite.PRepeat' 'Graphics.Vega.VegaLite.Row'@.
+
+-}
+data RepeatFields
+    = RowFields [T.Text]
+    | ColumnFields [T.Text]
+
+
+repeatFieldsProperty :: RepeatFields -> LabelledSpec
+repeatFieldsProperty rfs =
+  let (nme, vs) = case rfs of
+        RowFields fields -> ("row", fields)
+        ColumnFields fields -> ("column", fields)
+
+  in nme .= map toJSON vs
+
+
+{-|
+
+Indicates the type of color interpolation to apply, when mapping a data field
+onto a color scale.
+
+For details see the
+<https://vega.github.io/vega-lite/docs/scale.html#continuous Vega-Lite documentation>.
+
+-}
+data CInterpolate
+    = CubeHelix Double
+      -- ^ Cube helix color interpolation for continuous color scales using the given
+      --   gamma value (anchored at 1).
+    | CubeHelixLong Double
+      -- ^ Long-path cube helix color interpolation for continuous color scales using
+      --   the given gamma value (anchored at 1).
+    | Hcl
+      -- ^ HCL color interpolation for continuous color scales.
+    | HclLong
+      -- ^ HCL color interpolation in polar coordinate space for continuous color scales.
+    | Hsl
+      -- ^ HSL color interpolation for continuous color scales.
+    | HslLong
+      -- ^ HSL color interpolation in polar coordinate space for continuous color scales.
+    | Lab
+      -- ^ Lab color interpolation for continuous color scales.
+    | Rgb Double
+      -- ^ RGB color interpolation for continuous color scales using the given gamma
+      --   value (anchored at 1).
+
+
+-- Need to tie down some types as things are too polymorphic,
+-- particularly in the presence of OverloadedStrings.
+--
+pairT :: T.Text -> T.Text -> (T.Text, A.Value)
+pairT a b = a .= b
+
+
+cInterpolateSpec :: CInterpolate -> VLSpec
+cInterpolateSpec (Rgb gamma) = object [pairT "type" "rgb", "gamma" .= gamma]
+cInterpolateSpec Hsl = object [pairT "type" "hsl"]
+cInterpolateSpec HslLong = object [pairT "type" "hsl-long"]
+cInterpolateSpec Lab = object [pairT "type" "lab"]
+cInterpolateSpec Hcl = object [pairT "type" "hcl"]
+cInterpolateSpec HclLong = object [pairT "type" "hcl-long"]
+cInterpolateSpec (CubeHelix gamma) = object [pairT "type" "cubehelix", "gamma" .= gamma]
+cInterpolateSpec (CubeHelixLong gamma) = object [pairT "type" "cubehelix-long", "gamma" .= gamma]
